Move Tornado-specific handlers to tornado directory
